fix: Windows format specifiers in DHT code (v0.3.52) [BUILD]
- dht_bootstrap_registry.c: %lu -> %llu for uint64_t
- dht_context.cpp: Fix format specifiers for value_id (uint64_t)
  and duration counts (long long) on Windows

Remaining warnings are in vendor/opendht-pq (third-party code).

+++ dht/core/dht_bootstrap_registry.c
@@ -246,7 +246,7 @@ int dht_bootstrap_registry_register(
     uint64_t value_id = 1;
     dht_get_owner_value_id(dht_ctx, &value_id);
 
-    QGP_LOG_INFO(LOG_TAG, "Registering node_id in index (value_id=%lu)\n", value_id);
+    QGP_LOG_INFO(LOG_TAG, "Registering node_id in index (value_id=%llu)\n", (unsigned long long)value_id);
     ret = dht_put_signed(dht_ctx, (uint8_t*)index_key, strlen(index_key),
                          (uint8_t*)node_id, strlen(node_id),
                          value_id, DHT_CHUNK_TTL_7DAY);
